[Refactor/#135] SettingScreen, SettingViewModel을 screens.setting 패키지로 이동
